+++ .github/CHANGELOG.md
@@ -5,7 +5,7 @@
  * Add support for building X86-64 Linux wheels with OpenMP and SERIAL backends with Github Actions.
  [(#14)](https://github.com/PennyLaneAI/pennylane-lightning-kokkos/pull/14)
 
- * Add the `kokkos_config` class variable, which stores the kokkos build and runtime information such as `Backend`, `Arch`, `Kokkos Version`, `Runtime Config`, `Compiler`, to LightningKokkos for users' query purpose. Users can also access other information such as `Options`, `Memory`, `Atomics` and `Vectorization` from `kokkos_config`.   
+ * Add the `kokkos_config` class variable, which stores the kokkos build and runtime information such as `Backend`, `Architecture`, `Kokkos Version`, `Compiler`, to LightningKokkos for users' query purpose. Users can also access other information such as `Options`, `Memory`, `Atomics` and `Vectorization` from `kokkos_config`.   
  The workflow for build and runtime information query is:
  ```python
  >>> import pennylane as qml